+++ contracts/interfaces/IReadFacet.sol
@@ -6,17 +6,24 @@ import "./IRToken.sol";
 import "./IStRSR.sol";
 
 /**
- * @title IFacadeRead
+ * @title IReadFacet
  * @notice A UX-friendly layer for read operations, especially those that first require refresh()
  *
  * - @custom:static-call - Use ethers callStatic() in order to get result after update
 v */
-interface IFacadeRead {
+interface IReadFacet {
     // === Static Calls ===
 
     /// @return How many RToken `account` can issue given current holdings
     /// @custom:static-call
     function maxIssuable(IRToken rToken, address account) external returns (uint256);
+
+    /// @param amounts {qTok} The balances of each basket ERC20 to assume
+    /// @return How many RToken can be issued
+    /// @custom:static-call
+    function maxIssuableByAmounts(IRToken rToken, uint256[] memory amounts)
+        external
+        returns (uint256);
 
     /// @return tokens The erc20 needed for the issuance
     /// @return deposits {qTok} The deposits necessary to issue `amount` RToken