+++ scripts/verification/collateral-plugins/verify_convex_crvusd_usdc.ts
@@ -5,30 +5,24 @@ import { bn } from '../../../common/numbers'
 import { ONE_ADDRESS } from '../../../common/constants'
 import {
   getDeploymentFile,
-  getDeploymentFilename,
   getAssetCollDeploymentFilename,
   IAssetCollDeployments,
-  IDeployments,
 } from '../../deployment/common'
 import { verifyContract } from '../../deployment/utils'
 import { revenueHiding } from '../../deployment/utils'
 import {
   CurvePoolType,
-  DAI_ORACLE_ERROR,
-  DAI_ORACLE_TIMEOUT,
-  DAI_USD_FEED,
   DEFAULT_THRESHOLD,
   DELAY_UNTIL_DEFAULT,
   MAX_TRADE_VOL,
   PRICE_TIMEOUT,
-  THREE_POOL,
-  THREE_POOL_TOKEN,
+  crvUSD_USDC,
   USDC_ORACLE_ERROR,
   USDC_ORACLE_TIMEOUT,
   USDC_USD_FEED,
-  USDT_ORACLE_ERROR,
-  USDT_ORACLE_TIMEOUT,
-  USDT_USD_FEED,
+  crvUSD_ORACLE_ERROR,
+  crvUSD_ORACLE_TIMEOUT,
+  crvUSD_USD_FEED,
 } from '../../../test/plugins/individual-collateral/curve/constants'
 
 let deployments: IAssetCollDeployments
@@ -44,15 +38,12 @@ async function main() {
     throw new Error(`Cannot verify contracts for development chain ${hre.network.name}`)
   }
 
-  const deploymentFilename = getDeploymentFilename(chainId)
-  const coreDeployments = <IDeployments>getDeploymentFile(deploymentFilename)
-
   const assetCollDeploymentFilename = getAssetCollDeploymentFilename(chainId)
   deployments = <IAssetCollDeployments>getDeploymentFile(assetCollDeploymentFilename)
 
   const w3PoolCollateral = await ethers.getContractAt(
     'CurveStableCollateral',
-    deployments.collateral.cvx3Pool as string
+    deployments.collateral.cvxCrvUSDUSDC as string
   )
 
   /********  Verify ConvexStakingWrapper  **************************/
@@ -64,10 +55,10 @@ async function main() {
     'contracts/plugins/assets/curve/cvx/vendor/ConvexStakingWrapper.sol:ConvexStakingWrapper'
   )
 
-  /********  Verify 3Pool plugin  **************************/
+  /********  Verify crvUSD-USDC plugin  **************************/
   await verifyContract(
     chainId,
-    deployments.collateral.cvx3Pool,
+    deployments.collateral.cvxCrvUSDUSDC,
     [
       {
         erc20: await w3PoolCollateral.erc20(),
@@ -82,13 +73,13 @@ async function main() {
       },
       revenueHiding.toString(),
       {
-        nTokens: 3,
-        curvePool: THREE_POOL,
+        nTokens: 2,
+        curvePool: crvUSD_USDC,
         poolType: CurvePoolType.Plain,
-        feeds: [[DAI_USD_FEED], [USDC_USD_FEED], [USDT_USD_FEED]],
-        oracleTimeouts: [[DAI_ORACLE_TIMEOUT], [USDC_ORACLE_TIMEOUT], [USDT_ORACLE_TIMEOUT]],
-        oracleErrors: [[DAI_ORACLE_ERROR], [USDC_ORACLE_ERROR], [USDT_ORACLE_ERROR]],
-        lpToken: THREE_POOL_TOKEN,
+        feeds: [[USDC_USD_FEED], [crvUSD_USD_FEED]],
+        oracleTimeouts: [[USDC_ORACLE_TIMEOUT], [crvUSD_ORACLE_TIMEOUT]],
+        oracleErrors: [[USDC_ORACLE_ERROR], [crvUSD_ORACLE_ERROR]],
+        lpToken: crvUSD_USDC,
       },
     ],
     'contracts/plugins/assets/curve/CurveStableCollateral.sol:CurveStableCollateral'