+++ scripts/verify_etherscan.ts
@@ -49,7 +49,8 @@ async function main() {
   // Phase 2 - Individual Plugins
   if (!baseL2Chains.includes(hre.network.name)) {
     scripts.push(
-      'collateral-plugins/verify_convex_stable.ts',
+      'collateral-plugins/verify_convex_crvusd_usdc.ts',
+      'collateral-plugins/verify_convex_3pool.ts',
       'collateral-plugins/verify_convex_stable_metapool.ts',
       'collateral-plugins/verify_convex_stable_rtoken_metapool.ts',
       'collateral-plugins/verify_curve_stable.ts',
@@ -71,7 +72,7 @@ async function main() {
     scripts.push(
       'collateral-plugins/verify_cbeth.ts',
       'collateral-plugins/verify_cusdbcv3.ts',
-      'collateral-plugins/verify_aave_v3_usdbc',
+      'collateral-plugins/verify_aave_v3_usdc.ts',
       'collateral-plugins/verify_stargate_usdc',
       'assets/verify_stg.ts'
     )